fix: split hanlder to util files
