+++ src/cdn-analysis/handler.js
@@ -11,351 +11,19 @@
  */
 
 /* c8 ignore start */
-import { StartQueryExecutionCommand, GetQueryExecutionCommand, GetQueryResultsCommand } from '@aws-sdk/client-athena';
 import { AuditBuilder } from '../common/audit-builder.js';
-import { trafficAnalysisQueries } from './queries/traffic-analysis.js';
-import { referrerAnalysisQueries } from './queries/referrer-analysis.js';
-import { userAgentAnalysisQueries } from './queries/user-agent-analysis.js';
-import { errorAnalysisQueries } from './queries/error-analysis.js';
-import { geoAnalysisQueries } from './queries/geo-analysis.js';
-import { frequencyAnalysisQueries } from './queries/frequency-analysis.js';
 import { saveToS3AsParquet } from './aggregators/parquet-writer.js';
-import { getS3Config, getCustomerRawLogsLocation, getRawLogsPartitionConfig } from './config/s3-config.js';
+import { getS3Config } from './config/s3-config.js';
+import { ensureAthenaTablesExist, createFormattedLogsTable } from './utils/table-manager.js';
+import { runAllAnalysis, createAnalysisSummary } from './utils/analysis-runners.js';
+import { filterAndStoreAgenticLogs } from './utils/unload-operations.js';
 
 const INTERVAL = 1; // hours
 
 /**
- * Parse Athena results into usable format
+ * Step 1: Filter and store agentic logs
  */
-function parseAthenaResults(results) {
-  if (!results.ResultSet || !results.ResultSet.Rows || results.ResultSet.Rows.length === 0) {
-    return [];
-  }
-
-  const rows = results.ResultSet.Rows;
-  const headers = rows[0].Data.map((col) => col.VarCharValue);
-
-  return rows.slice(1).map((row) => {
-    const record = {};
-    row.Data.forEach((col, index) => {
-      record[headers[index]] = col.VarCharValue;
-    });
-    return record;
-  });
-}
-
-/**
- * Wait for Athena query execution to complete
- */
-async function waitForQueryExecution(athenaClient, queryExecutionId, maxAttempts = 60) {
-  let queryExecution;
-  let attempts = 0;
-
-  // eslint-disable-next-line no-await-in-loop
-  while (attempts < maxAttempts) {
-    // eslint-disable-next-line no-await-in-loop, no-promise-executor-return
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-    const getCommand = new GetQueryExecutionCommand({ QueryExecutionId: queryExecutionId });
-    // eslint-disable-next-line no-await-in-loop
-    queryExecution = await athenaClient.send(getCommand);
-    attempts += 1;
-
-    const state = queryExecution.QueryExecution.Status.State;
-    if (state !== 'RUNNING' && state !== 'QUEUED') {
-      break;
-    }
-  }
-
-  // If we've reached max attempts and query is still running, throw an error
-  const finalState = queryExecution.QueryExecution.Status.State;
-  if ((finalState === 'RUNNING' || finalState === 'QUEUED') && attempts >= maxAttempts) {
-    throw new Error(`Query execution timed out after ${maxAttempts} attempts. Current state: ${finalState}`);
-  }
-
-  return queryExecution;
-}
-
-/**
- * Execute a setup query and wait for completion
- */
-async function executeAthenaSetupQuery(athenaClient, query, description, s3Config, log) {
-  try {
-    log.info(`🔧 Setting up ${description}...`);
-
-    const startCommand = new StartQueryExecutionCommand({
-      QueryString: query,
-      QueryExecutionContext: { Database: 'default' },
-      ResultConfiguration: {
-        OutputLocation: s3Config.getAthenaTempLocation(),
-      },
-    });
-
-    const startResult = await athenaClient.send(startCommand);
-    const queryExecutionId = startResult.QueryExecutionId;
-
-    // Wait for completion
-    const queryExecution = await waitForQueryExecution(athenaClient, queryExecutionId);
-
-    if (queryExecution.QueryExecution.Status.State === 'SUCCEEDED') {
-      log.info(`✅ ${description} setup completed`);
-    } else {
-      const error = queryExecution.QueryExecution.Status.StateChangeReason || 'Unknown error';
-      log.warn(`⚠️ ${description} setup issue: ${error}`);
-      // Don't throw for table creation issues - might already exist
-    }
-  } catch (error) {
-    log.warn(`⚠️ ${description} setup warning:`, error.message);
-    // Don't throw - continue with analysis even if setup has issues
-  }
-}
-
-/**
- * Execute Athena query and return results
- */
-async function executeAthenaQuery(athenaClient, query, s3Config, log, database = 'cdn_logs') {
-  try {
-    // Start query execution
-    const startCommand = new StartQueryExecutionCommand({
-      QueryString: query,
-      QueryExecutionContext: { Database: database },
-      ResultConfiguration: {
-        OutputLocation: s3Config.getAthenaTempLocation(),
-      },
-    });
-
-    const startResult = await athenaClient.send(startCommand);
-    const queryExecutionId = startResult.QueryExecutionId;
-
-    // Wait for query completion
-    const queryExecution = await waitForQueryExecution(athenaClient, queryExecutionId);
-
-    if (queryExecution.QueryExecution.Status.State !== 'SUCCEEDED') {
-      throw new Error(`Query failed: ${queryExecution.QueryExecution.Status.StateChangeReason}`);
-    }
-
-    // Get query results
-    const resultsCommand = new GetQueryResultsCommand({ QueryExecutionId: queryExecutionId });
-    const results = await athenaClient.send(resultsCommand);
-
-    return parseAthenaResults(results);
-  } catch (error) {
-    log.error('Athena query execution failed:', error);
-    throw error;
-  }
-}
-
-/**
- * Run traffic analysis
- */
-async function runTrafficAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log) {
-  const query = trafficAnalysisQueries.hourlyTraffic(hourToProcess, customerTableName);
-  return executeAthenaQuery(athenaClient, query, s3Config, log, 'cdn_logs');
-}
-
-/**
- * Run referrer analysis
- */
-async function runReferrerAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log) {
-  const query = referrerAnalysisQueries.hourlyReferrers(hourToProcess, customerTableName);
-  return executeAthenaQuery(athenaClient, query, s3Config, log, 'cdn_logs');
-}
-
-/**
- * Run user agent analysis
- */
-async function runUserAgentAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log) {
-  const query = userAgentAnalysisQueries.hourlyUserAgents(hourToProcess, customerTableName);
-  return executeAthenaQuery(athenaClient, query, s3Config, log, 'cdn_logs');
-}
-
-/**
- * Run error analysis
- */
-async function runErrorAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log) {
-  const query = errorAnalysisQueries.hourlyErrors(hourToProcess, customerTableName);
-  return executeAthenaQuery(athenaClient, query, s3Config, log, 'cdn_logs');
-}
-
-/**
- * Run geographic analysis
- */
-async function runGeoAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log) {
-  const query = geoAnalysisQueries.hourlyByCountry(hourToProcess, customerTableName);
-  return executeAthenaQuery(athenaClient, query, s3Config, log, 'cdn_logs');
-}
-
-/**
- * Run frequency analysis
- */
-async function runFrequencyAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log) {
-  const query = frequencyAnalysisQueries.hourlyFrequencyPatterns(hourToProcess, customerTableName);
-  return executeAthenaQuery(athenaClient, query, s3Config, log, 'cdn_logs');
-}
-
-/**
- * Create summary of all analysis results
- */
-function createAnalysisSummary(analysisResults, hourProcessed, s3Config) {
-  const summary = {
-    timestamp: new Date().toISOString(),
-    hourProcessed: hourProcessed.toISOString(),
-    customerDomain: s3Config.customerDomain,
-    environment: s3Config.environment,
-    analysisTypes: Object.keys(analysisResults),
-    recordCounts: {},
-    totalRequests: 0,
-    agenticRequests: 0,
-  };
-
-  // Count records in each analysis
-  Object.entries(analysisResults).forEach(([type, data]) => {
-    summary.recordCounts[type] = Array.isArray(data) ? data.length : 0;
-  });
-
-  // Extract key metrics
-  if (analysisResults.traffic && analysisResults.traffic.length > 0) {
-    summary.totalRequests = parseInt(analysisResults.traffic[0].total_requests || 0, 10);
-  }
-
-  if (analysisResults.userAgent) {
-    summary.agenticRequests = analysisResults.userAgent
-      .filter((ua) => ua.is_agentic === 'true')
-      .reduce((sum, ua) => sum + parseInt(ua.count || 0, 10), 0);
-  }
-
-  summary.agenticPercentage = summary.totalRequests > 0
-    ? (summary.agenticRequests / summary.totalRequests) * 100
-    : 0;
-
-  return summary;
-}
-
-/**
- * Ensure Athena tables exist, create them if they don't
- */
-async function ensureAthenaTablesExist(athenaClient, s3Config, log) {
-  try {
-    log.info('🔧 Checking Athena tables setup...');
-
-    // Create database first
-    await executeAthenaSetupQuery(
-      athenaClient,
-      'CREATE DATABASE IF NOT EXISTS cdn_logs',
-      'cdn_logs database',
-      s3Config,
-      log,
-    );
-
-    // Create customer-specific raw logs table
-    const rawLogsLocation = getCustomerRawLogsLocation(s3Config);
-    const customerTableName = `raw_logs_${s3Config.customerDomain}`;
-    const partitionConfig = getRawLogsPartitionConfig(s3Config);
-
-    const rawLogsTableDDL = `
-      CREATE EXTERNAL TABLE IF NOT EXISTS cdn_logs.${customerTableName} (
-        timestamp string,
-        geo_country string,
-        host string,
-        url string,
-        request_method string,
-        request_protocol string,
-        request_user_agent string,
-        response_state string,
-        response_status int,
-        response_reason string,
-        referer string
-      )
-      PARTITIONED BY (
-        year string,
-        month string,
-        day string,
-        hour string
-      )
-      ROW FORMAT SERDE 'org.apache.hive.hcatalog.data.JsonSerDe'
-      LOCATION '${rawLogsLocation}'
-      TBLPROPERTIES (
-        'projection.enabled' = '${partitionConfig.projectionEnabled}',
-        'storage.location.template' = '${partitionConfig.locationTemplate}',
-        ${Object.entries(partitionConfig.partitionProjections).map(([key, value]) => `'${key}' = '${value}'`).join(',\n        ')},
-        'has_encrypted_data' = 'false'
-      )
-    `;
-
-    await executeAthenaSetupQuery(
-      athenaClient,
-      rawLogsTableDDL,
-      `${customerTableName} table`,
-      s3Config,
-      log,
-    );
-
-    // Create new Parquet analysis table for improved performance
-    const parquetAnalysisTableDDL = `
-      CREATE EXTERNAL TABLE IF NOT EXISTS cdn_logs.cdn_analysis_data (
-        analysis_type string,
-        customer_domain string,
-        hour_processed timestamp,
-        generated_at timestamp,
-        record_index int,
-        total_requests bigint,
-        success_rate double,
-        agentic_requests bigint,
-        geo_country string,
-        response_status bigint,
-        request_user_agent string,
-        referer string,
-        additional_data string
-      )
-      PARTITIONED BY (
-        customer string,
-        year string,
-        month string,
-        day string,
-        hour string
-      )
-      STORED AS PARQUET
-      LOCATION '${s3Config.analysisBucket}/cdn-analysis/'
-      TBLPROPERTIES (
-        'projection.enabled' = 'true',
-        'projection.customer.type' = 'enum',
-        'projection.customer.values' = 'blog_adobe_com,other_domains',
-        'projection.year.type' = 'integer',
-        'projection.year.range' = '2024,2030',
-        'projection.year.format' = 'yyyy',
-        'projection.month.type' = 'integer',
-        'projection.month.range' = '1,12',
-        'projection.month.format' = 'MM',
-        'projection.day.type' = 'integer',
-        'projection.day.range' = '1,31',
-        'projection.day.format' = 'dd',
-        'projection.hour.type' = 'integer',
-        'projection.hour.range' = '0,23',
-        'projection.hour.format' = 'HH',
-        'storage.location.template' = 's3://${s3Config.analysisBucket}/cdn-analysis/customer=\${customer}/year=\${year}/month=\${month}/day=\${day}/hour=\${hour}/',
-        'has_encrypted_data' = 'false'
-      )
-    `;
-
-    await executeAthenaSetupQuery(
-      athenaClient,
-      parquetAnalysisTableDDL,
-      'cdn_analysis_parquet table',
-      s3Config,
-      log,
-    );
-
-    log.info('✅ Athena tables setup completed successfully!');
-  } catch (error) {
-    log.error('❌ Failed to setup Athena tables:', error);
-    throw error;
-  }
-}
-
-/**
- * CDN Analysis Runner - Comprehensive CDN log analysis
- * Supports 6 analysis types: Traffic, Referrer, User Agent, Error, Geographic, Frequency
- */
-export async function CDNAnalysisRunner(auditUrl, context, site) {
+async function filterAgenticLogsStep(auditUrl, context, site) {
   const { log, athenaClient, s3Client } = context;
 
   if (!athenaClient) {
@@ -373,40 +41,75 @@ export async function CDNAnalysisRunner(auditUrl, context, site) {
   const now = new Date();
   const hourToProcess = new Date(now.getTime() - (60 * 60 * 1000)); // 1 hour ago
 
-  log.info(`Starting CDN analysis for ${auditUrl} - processing hour: ${hourToProcess.toISOString()}`);
-
-  const analysisResults = {};
+  log.info(`Step 1: Filtering agentic logs for ${auditUrl} - processing hour: ${hourToProcess.toISOString()}`);
 
   try {
     // Auto-setup: Ensure Athena tables exist
     await ensureAthenaTablesExist(athenaClient, s3Config, log);
-    // Run all 6 analysis types in parallel
-    const customerTableName = `raw_logs_${s3Config.customerDomain}`;
-    const analysisPromises = [
-      runTrafficAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log),
-      runReferrerAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log),
-      runUserAgentAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log),
-      runErrorAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log),
-      runGeoAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log),
-      runFrequencyAnalysis(athenaClient, hourToProcess, s3Config, customerTableName, log),
-    ];
 
-    const [
-      trafficData,
-      referrerData,
-      userAgentData,
-      errorData,
-      geoData,
-      frequencyData,
-    ] = await Promise.allSettled(analysisPromises);
+    // Note: Using UNLOAD to export directly to S3, no table needed
 
-    // Collect successful results
-    if (trafficData.status === 'fulfilled') analysisResults.traffic = trafficData.value;
-    if (referrerData.status === 'fulfilled') analysisResults.referrer = referrerData.value;
-    if (userAgentData.status === 'fulfilled') analysisResults.userAgent = userAgentData.value;
-    if (errorData.status === 'fulfilled') analysisResults.error = errorData.value;
-    if (geoData.status === 'fulfilled') analysisResults.geo = geoData.value;
-    if (frequencyData.status === 'fulfilled') analysisResults.frequency = frequencyData.value;
+    // Filter and store agentic logs
+    const sourceTableName = `raw_logs_${s3Config.customerDomain}`;
+    const agenticLogsCount = await filterAndStoreAgenticLogs(
+      athenaClient,
+      hourToProcess,
+      s3Config,
+      sourceTableName,
+      log,
+    );
+
+    log.info(`✅ Step 1 completed: Filtered and stored ${agenticLogsCount} agentic log entries`);
+
+    return {
+      auditResult: {
+        step: 'filter-agentic-logs',
+        hourProcessed: hourToProcess.toISOString(),
+        agenticLogsCount,
+        unloadMethod: 'UNLOAD_TO_S3',
+        storagePath: `s3://${s3Config.rawLogsBucket}/formatted-logs/`,
+        format: 'JSON_GZIPPED',
+        customerDomain: s3Config.customerDomain,
+        environment: s3Config.environment,
+      },
+      fullAuditRef: auditUrl,
+    };
+  } catch (error) {
+    log.error(`Step 1 failed for ${auditUrl}:`, error);
+    throw error;
+  }
+}
+
+/**
+ * Step 2: Run aggregated analysis on filtered agentic logs
+ */
+async function runAggregatedAnalysisStep(auditUrl, context, site) {
+  const { log, athenaClient, s3Client } = context;
+
+  // Get S3 configuration based on customer and environment
+  const s3Config = getS3Config(context, site);
+
+  // Get the hour to process (previous hour)
+  const now = new Date();
+  const hourToProcess = new Date(now.getTime() - (60 * 60 * 1000)); // 1 hour ago
+
+  log.info(`Step 2: Running aggregated analysis for ${auditUrl} - processing hour: ${hourToProcess.toISOString()}`);
+
+  try {
+    // Create table on-demand to read from UNLOAD files
+    await createFormattedLogsTable(athenaClient, s3Config, log);
+
+    // Use formatted logs table for analysis
+    const formattedTableName = `formatted_logs_${s3Config.customerDomain}`;
+
+    // Run all analysis types in parallel using utility function
+    const analysisResults = await runAllAnalysis(
+      athenaClient,
+      hourToProcess,
+      s3Config,
+      formattedTableName,
+      log,
+    );
 
     // Save all results to S3 as Parquet files
     const s3SavePromises = Object.entries(analysisResults)
@@ -415,7 +118,7 @@ export async function CDNAnalysisRunner(auditUrl, context, site) {
         data,
         hourProcessed: hourToProcess,
         bucket: s3Config.analysisBucket,
-        basePrefix: 'cdn-analysis',
+        basePrefix: 'cdn-analysis-agentic',
         log,
         customerDomain: s3Config.customerDomain,
         s3Client,
@@ -426,13 +129,15 @@ export async function CDNAnalysisRunner(auditUrl, context, site) {
     // Create summary
     const summary = createAnalysisSummary(analysisResults, hourToProcess, s3Config);
 
-    const logMessage = `CDN analysis completed for ${auditUrl} (customer: ${s3Config.customerDomain}).`
-      + ` Processed ${Object.keys(analysisResults).length} analysis types.`;
+    const logMessage = `Step 2 completed: CDN agentic analysis for ${auditUrl} `
+      + `(customer: ${s3Config.customerDomain}). `
+      + `Processed ${Object.keys(analysisResults).length} analysis types.`;
     log.info(logMessage);
 
     return {
       auditResult: {
-        cdnAnalysis: summary,
+        step: 'aggregated-analysis',
+        cdnAgenticAnalysis: summary,
         auditContext: {
           interval: INTERVAL,
           hourProcessed: hourToProcess.toISOString(),
@@ -444,12 +149,54 @@ export async function CDNAnalysisRunner(auditUrl, context, site) {
       fullAuditRef: auditUrl,
     };
   } catch (error) {
-    log.error(`CDN analysis failed for ${auditUrl}:`, error);
+    log.error(`Step 2 failed for ${auditUrl}:`, error);
+    throw error;
+  }
+}
+
+/**
+ * Main CDN Analysis Runner - executes both steps sequentially
+ */
+async function runCdnAnalysis(auditUrl, context) {
+  const { log } = context;
+
+  log.info(`Starting CDN Analysis for ${auditUrl}`);
+
+  try {
+    // Get site information
+    const site = context.site || { baseURL: auditUrl };
+
+    // Step 1: Filter and store agentic logs
+    log.info('Executing Step 1: Filter agentic logs');
+    const step1Result = await filterAgenticLogsStep(auditUrl, context, site);
+
+    // Step 2: Run aggregated analysis
+    log.info('Executing Step 2: Run aggregated analysis');
+    const step2Result = await runAggregatedAnalysisStep(auditUrl, context, site);
+
+    // Combine results
+    const combinedResult = {
+      auditResult: {
+        step1: step1Result.auditResult,
+        step2: step2Result.auditResult,
+        summary: {
+          totalSteps: 2,
+          completedAt: new Date().toISOString(),
+          auditUrl,
+        },
+      },
+      fullAuditRef: auditUrl,
+    };
+
+    log.info(`✅ CDN Analysis completed successfully for ${auditUrl}`);
+    return combinedResult;
+  } catch (error) {
+    log.error(`❌ CDN Analysis failed for ${auditUrl}:`, error);
     throw error;
   }
 }
 
 export default new AuditBuilder()
-  .withRunner(CDNAnalysisRunner)
+  .withRunner(runCdnAnalysis)
   .build();
 /* c8 ignore stop */
